requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.2
google-auth==2.40.2
google-genai==1.17.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.7
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
sniffio==1.3.1
textual==3.1.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uc-micro-py==1.0.3
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
